Likelihood and Impact added
Arindam had the issue where he needed to add an extra row to the csv files with an extra value. before it was just substituting an X with a value, but now it is adding a second row with all empty commas except the likehood values which will be taken from a second column.